Disable GUIEnter autocommand
It's nearly absent from the diff, but there's a good three paragraphs of
comments above this position in the code that explains this change.

In fact these lines should never have been written in the first place.
This color scheme used to only specify EITHER the cterm OR the gui
colors. Colorschemes should always specify both.

We refactored this correctly when we updated the colorscheme to use
24-bit colors.